let's keep operator-specific additions in a separate directory
