+++ src/frontend/org/voltdb/RealVoltDB.java
@@ -176,6 +176,10 @@ import org.voltdb.licensetool.LicenseApi;
 import org.voltdb.messaging.MigratePartitionLeaderMessage;
 import org.voltdb.messaging.VoltDbMessageFactory;
 import org.voltdb.modular.ModuleManager;
+import org.voltdb.oper.PauseActivityStats;
+import org.voltdb.oper.ShutdownActivityStats;
+import org.voltdb.oper.StatusListener;
+import org.voltdb.oper.StopActivityStats;
 import org.voltdb.planner.ActivePlanRepository;
 import org.voltdb.probe.MeshProber;
 import org.voltdb.processtools.ShellTools;
@@ -209,7 +213,6 @@ import org.voltdb.utils.LogKeys;
 import org.voltdb.utils.MiscUtils;
 import org.voltdb.utils.PlatformProperties;
 import org.voltdb.utils.ProClass;
-import org.voltdb.utils.StatusListener;
 import org.voltdb.utils.SystemStatsCollector;
 import org.voltdb.utils.TopologyZKUtils;
 import org.voltdb.utils.VoltFile;
