+++ src/frontend/org/voltdb/oper/PauseActivityStats.java
@@ -15,19 +15,22 @@
  * along with VoltDB.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.voltdb;
+package org.voltdb.oper;
 
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.Set;
 
+import org.voltdb.StatsSelector;
+import org.voltdb.StatsSource;
 import org.voltdb.VoltTable.ColumnInfo;
+import org.voltdb.VoltType;
 import org.voltcore.logging.VoltLogger;
 
 /**
- * The StopActivityStats statistics provide a summary of current
+ * The PauseActivityStats statistics provide a summary of current
  * cluster activity that can be used to determine when the cluster
- * activity has subsided after a call to @PrepareStopNode.
+ * activity has subsided after a call to @Pause.
  *
  * The result of the statistics request is a table with one
  * row per host, and a summary of activity in various categories.
@@ -35,16 +38,16 @@ import org.voltcore.logging.VoltLogger;
  * then other columns can be used to determine what the activity
  * relates to, and perhaps whether forward progress is being made.
  */
-public class StopActivityStats extends StatsSource {
+public class PauseActivityStats extends StatsSource {
     private static final VoltLogger logger = new VoltLogger("HOST");
 
     private enum ColumnName {
         ACTIVE, // 0 if all other gauges 0, else 1
-        PAR_LEADERS,
-        EXP_MASTERS,
+        EXPORTS_PENDING,
+        DRPROD_ROWS, DRPROD_BYTES,
     };
 
-    public StopActivityStats() {
+    public PauseActivityStats() {
         super(false);
     }
 
@@ -75,7 +78,7 @@ public class StopActivityStats extends StatsSource {
      * used by 'voltadmin pause --wait'.
      */
     private static final ActivityHelper.Type[] statsList = {
-        ActivityHelper.Type.LEADER, ActivityHelper.Type.EXMAST
+        ActivityHelper.Type.EXPORT, ActivityHelper.Type.DRPROD,
     };
 
     @Override
@@ -84,11 +87,12 @@ public class StopActivityStats extends StatsSource {
         try {
             ActivityHelper helper = new ActivityHelper();
             active = helper.collect(statsList);
-            setValue(row, ColumnName.PAR_LEADERS, helper.leaderCount);
-            setValue(row, ColumnName.EXP_MASTERS, helper.exportMasters);
+            setValue(row, ColumnName.EXPORTS_PENDING, helper.exportPend);
+            setValue(row, ColumnName.DRPROD_ROWS, helper.drprodRowsPend);
+            setValue(row, ColumnName.DRPROD_BYTES, helper.drprodBytesPend);
         }
         catch (Exception ex) {
-            logger.error("Unhandled exception in StopActivityStats: " + ex);
+            logger.error("Unhandled exception in PauseActivityStats: " + ex);
         }
         setValue(row, ColumnName.ACTIVE, active);
         super.updateStatsRow(key, row);
