+++ src/frontend/org/voltdb/oper/StatusListener.java
@@ -15,7 +15,7 @@
  * along with VoltDB.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.voltdb.utils;
+package org.voltdb.oper;
 
 import java.io.IOException;
 import java.net.InetAddress;
