+++ src/frontend/org/voltdb/oper/StatusServlet.java
@@ -15,7 +15,7 @@
  * along with VoltDB.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.voltdb.utils;
+package org.voltdb.oper;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -23,6 +23,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.json_voltpatches.JSONObject;
+
 import org.voltdb.VoltDB;
 import org.voltdb.VoltDBInterface;
